refactor(cli): split into package
