+++ pyproject.toml
@@ -49,7 +49,7 @@ dev = [
 ]
 
 [project.scripts]
-udata-hydra = "udata_hydra.cli_entry:run"
+udata-hydra = "udata_hydra.cli:run"
 udata-hydra-crawl = "udata_hydra.crawl:run"
 udata-hydra-app = "udata_hydra.app:run"
