+++ tests/test_cli/test_probe_cors.py
@@ -17,7 +17,7 @@ async def test_probe_cors_cli_logs_result(monkeypatch, caplog):
         assert target == url
         return expected
 
-    monkeypatch.setattr("udata_hydra.cli.probe_cors", fake_probe)
+    monkeypatch.setattr("udata_hydra.cli.crawl.probe_cors", fake_probe)
     caplog.set_level(logging.INFO)
 
     await probe_cors_cli(url=url, resource_id=None)
